refactor(common): establish transaction type domain owner
